releasetools: Generate recovery-from-boot.p with bsdiff if necessary
If TARGET_NOT_USE_GZIP_RECOVERY_RAMDISK is true, we want to use bsdiff
to create recovery-from-boot.p, otherwise there are high chances that
imgdiff will fail. Currently this is done only when running make, but
not when we re-create the patch from releasetools, so do it.

Also, since recovery-resource.dat is used as bonus data for imgdiff,
don't build it if we are going to use bsdiff.

Change-Id: I93a662a358ee79f56b8acd4329eedd166a176c66
(cherry picked from commit 543d7376be045ccdfb20cc14b3757500a026e326)
Signed-off-by: Marco Zanin (B--B) <mrczn.bb@gmail.com>
Signed-off-by: mohancm100 <mohanmanjappa@gmail.com>